requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg==3.2.6
python-dotenv==1.0.1
drf-spectacular~=0.28.0